This is a first pass at creating feature transformation extractors. This includes support for Zero-Mean-Unit-Stddev and Min-Max Normalization as well as TF*IDF scaling.
There is a rudimentary example that shows how one would write the training data in such a scenario.  This example still needs to be fleshed out, and when it is done should also provide an example of the evaluation infrastructure.  To support this data, I have added data from the 20 newsgroups corpus - but in this case it is only a small sample of files from two newsgroups (hence the directory 2newsgroups).

There are also basic tests in place for a couple of the components, but this is still a large to do.

git-svn-id: https://cleartk.googlecode.com/svn/trunk@3750 2f4ce1e4-9571-11dd-93b2-6d14b300b761